fix(session replay): move removeInvalidSelectorsFromPrivacyConfig to joined config class
