+++ packages/session-replay-browser/src/config/joined-config.ts
@@ -8,7 +8,35 @@ import {
   SessionReplayRemoteConfig,
 } from './types';
 import { getDebugConfig } from '../helpers';
+import { Logger } from '@amplitude/analytics-types';
 
+export const removeInvalidSelectorsFromPrivacyConfig = (privacyConfig: PrivacyConfig, loggerProvider: Logger) => {
+  // This allows us to not search the DOM.
+  const fragment = document.createDocumentFragment();
+
+  const dropInvalidSelectors = (selectors: string[] | string = []): string[] | undefined => {
+    if (typeof selectors === 'string') {
+      selectors = [selectors];
+    }
+    selectors = selectors.filter((selector: string) => {
+      try {
+        fragment.querySelector(selector);
+      } catch {
+        loggerProvider.warn(`[session-replay-browser] omitting selector "${selector}" because it is invalid`);
+        return false;
+      }
+      return true;
+    });
+    if (selectors.length === 0) {
+      return undefined;
+    }
+    return selectors;
+  };
+  privacyConfig.blockSelector = dropInvalidSelectors(privacyConfig.blockSelector);
+  privacyConfig.maskSelector = dropInvalidSelectors(privacyConfig.maskSelector);
+  privacyConfig.unmaskSelector = dropInvalidSelectors(privacyConfig.unmaskSelector);
+  return privacyConfig;
+};
 export class SessionReplayJoinedConfigGenerator {
   localConfig: ISessionReplayLocalConfig;
   remoteConfigFetch: RemoteConfigFetch<SessionReplayRemoteConfig> | undefined;
@@ -154,7 +182,10 @@ export class SessionReplayJoinedConfigGenerator {
         }
       }
 
-      config.privacyConfig = joinedPrivacyConfig;
+      config.privacyConfig = removeInvalidSelectorsFromPrivacyConfig(
+        joinedPrivacyConfig,
+        this.localConfig.loggerProvider,
+      );
     }
 
     this.localConfig.loggerProvider.debug(
